Initial import of all RDBE+FieldSystem stuff
Received a lot of material from many people (thanks Uwe, Luis, Ed!) and
saving it here on private github first.

Now we can manage the code a bit better and also it is not only living on my
notebook so loss of that hardware won't be a disaster for *this* stuff.