Decomposing seconds to y,d,h,m,s procesing!
Noticed: used the same loop > 1 with only very slight differences ...
Bascially decomposing a time stamp in seconds into #-of-years, #-of-days,
hours, minutes and seconds and then handling the successive fields only
slightly differently.

So.

Decomposing a number of seconds into successively finer grained components
is already a reduce operation; each step of the reduction does a divmod()
with a different divisor and yielding the current number of wholes and a
remainder. The remainder will be the new "accumulation value" so the next
step of the reduction can take out a number of smaller units out of that
etc.

Now there is machinery which does the decomposing for you. You supply a list
of functions that will be called with the matching fields's value. Also you
supply a reductor which will reduce all the collected transformed components
into a pair of (remainder, result).
"remainder" is the number of fractional seconds. For integer second time
stamps this should be zero (and is enforced by some functions)

Function:
    decompose_seconds(reductor, actions) =>
        returns function with calling convention
             "(seconds, initial_accumulator)"
    reductor = function taking "(accumulator, result-of-action-on-field)"
        which should return a new accumulator value

Example:
   Desire: value in seconds to VEX time string "0000y000d00h00m00s"

   'actions' is a list of functions, each transforming the successive
   fields ('y','d','h','m','s') into appropriate strings, e.g.:

    actions = [  "{0:04d}y".format,  # convert nr of years to "0000y"
                 "{0:03d}d".format,  #      ..       days  to "000d"
                  ... etc ...
                 "{0:02d}s" ]        # the last field, seconds

    The 'reductor' can be just "operator.add" - adding strings together
    will append them into a single string. Thus:

      vextime_maker = decompose_seconds(operator.add, actions)

    Now 'vextime_maker' can be called like this:

        remain, result = vextime_maker(seconds, "")

    We pass the empty string as initial accumulator so 'result'
    will just contain the stringified time stamp. If a non-empty string
    was given, the string representation of the time will be appended
    to that string, obviously.

    'remain' is the fractional seconds remaining, which may or may not be of
    interest to you

    The whole thing can be simplified (empty string, discard remainder) like
    this:

    seconds2vex = lambda seconds: vextime_maker(seconds, "")[1]

In the code the exact same machinery is used to convert a number of seconds
into a proper VEXTime object - it has attributes 'y', 'd', 'h', 'm', 's' -
by having the actions convert the fields to tuples
    "(attribute-name, attribute-value)"
and letting the reductor set attribute value on the accumulator, which is
an empty VEXTime() object ...