Add logging to rdbe_sys_receive, cleanup
The Makefile doesn't build rdbe_tsys_receive_[01] anymore and the source
files deleted.

rdbe_tsys_receive has a built-in logging facility now. The Makefile also
creates a symlink "/usr2/st/bin/rdbe_tsys_receive_log" which points the
actual binary. If the binary detects it is being called through the symlink
(or basically any link ending in "_log") then it opens a file
    /tmp/rdbe_tsys_{rdbe}.log
and outputs decoded information from the Tsys broadcasts received.

This can be helpful in monitoring and/or debugging the system without
clobbering the FS terminal output.