Don't send "dbe_execute=init" after config
So part of the FPGA was already being configured between loading of the
personality and dbe_execute=init, possibly totally erasing the work done
before.
Also, because the personality is bluntly loaded, there is no need for this
anymore.

Secondly, the "rdbe_send=on/off" calls were removed because each setup was
already surrounded by a global "stop_send" and an "enable send for the whole
year" for the DBEs used in the setup.