Add /etc/{init.d,default}/fs-github-mount
These bootscripts allow you to remount the correct subdirectories in a git-controlled
directory containing a checkout of https://github.com/RDBE onto entries
/usr2/{st,control,proc}_github such that /usr2/{st,control,proc} can be
symlinked to those AND that everything (=the FS) still works.

The problem is that if /usr2/st is symlinked to *anything* in a different
location than /usr2/, it is impossible to compile the station code.

This fix introduces trampolines /usr2/{st,control,proc}_github which are
remounts of a another part of the file system tree using a bind mount
    https://unix.stackexchange.com/questions/198590/what-is-a-bind-mount
(trick courtesy of Arun Venkatamaran, AO, THANKS!)

After cloning the repo copy etc/{init.d,default}/fs-github-mount to
/etc/{...}/ on the FS computer.

Edit /etc/default/fs-github-mount to point at the directory where the git
clone actually is.