Add eVLBI recorder + set_disks fix + smarter
Added the "evlbi" recorder as supported type.
It will issue "net2mem=open"/"net2mem=close" for record=on/off.

The default Mark6 disks for type Mark6 should be "mk6" not "mark6"; the
jive5ab alias for "all Mark6 disks" is "mk6".

The "checkmk5" is only put in the .snp if the recorder actually (attempted)
to record anything; recorders "None" and "eVLBI" don't actually put anything
on disk so checking is pointless.